Prepare for Godliness--cleanliness is next to Godliness, right?
This is the final clean up; the last time I burn everything down to start
'afresh.' Everything is going to be built with (code) cleanliness in mind,
meaning: ugly code, code that has too many hacks and warts, will not be
accepted into a 'stable' release from here on.

You'll likely notice that the previous hack, a well-intented hotfix for
maintaining D.R.Y. has been replaced in favour of the Ryb::Gem module.
This means that the Gemspec reads from the (mostly) dependency-less and
portable (plain ol' Ruby) module and not the other way around--NICE.

This is the base-level of quality I (and hopefully at some point: we) will
strive for in Ryb (and surrounding support libraries).